separate domain and data component to remote library
