+++ README.md
@@ -8,9 +8,9 @@
 
 ## 公告：
 
-### 2020.10
+### 2020.11
 
-很高兴有位小伙伴分享了[面试心得](https://juejin.im/pin/5f1a889151882551517ff9b2)。截至目前，专栏已对体系化文章做了 830 余次修订，对专栏感兴趣的小伙伴可借助 [这份在 GitBook 上维护的 “导读”](https://kunminx.gitbook.io/relearn-android) 来获取它的目录 以及了解它的最新动态和发展状况。
+很高兴有位小伙伴分享了[面试心得](https://juejin.im/pin/5f1a889151882551517ff9b2)。截至目前，专栏已对体系化文章做了 870 余次修订，对专栏感兴趣的小伙伴可借助 [这份在 GitBook 上维护的 “导读”](https://kunminx.gitbook.io/relearn-android) 来获取它的目录 以及了解它的最新动态和发展状况。
 
 ### 2020.5
