+++ app/src/main/java/com/kunminx/puremusic/domain/usecase/CanBeStoppedUseCase.java
@@ -22,7 +22,7 @@ import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.kunminx.architecture.data.repository.DataResult;
+import com.kunminx.architecture.data.response.DataResult;
 import com.kunminx.architecture.domain.usecase.UseCase;
 import com.kunminx.puremusic.data.bean.DownloadFile;
 import com.kunminx.puremusic.data.repository.DataRepository;
