+++ architecture/build.gradle
@@ -57,4 +57,7 @@ dependencies {
     api 'com.kunminx.archi:unpeek-livedata:4.4.1-beta1'
     api 'com.kunminx.archi:smooth-navigation:3.3.2-beta5'
     api 'com.kunminx.archi:strict-databinding:3.5.1-beta1'
+    api 'com.kunminx.archi:domain:1.0.1'
+    api 'com.kunminx.archi:data:1.0.1'
+
 }
