+++ architecture/src/main/java/com/kunminx/architecture/data/response/manager/NetworkStateReceive.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.kunminx.architecture.domain.manager;
+package com.kunminx.architecture.data.response.manager;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
