--- architecture/src/main/java/com/kunminx/architecture/domain/manager/NetState.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2018-present KunMinX
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.kunminx.architecture.domain.manager;
-
-/**
- * Create by KunMinX at 19/10/11
- */
-public class NetState {
-
-    private String responseCode;
-    private boolean success = true;
-
-    public NetState(String responseCode, boolean success) {
-        this.responseCode = responseCode;
-        this.success = success;
-    }
-
-    public NetState() {
-    }
-
-    public String getResponseCode() {
-        return responseCode;
-    }
-
-    public void setResponseCode(String responseCode) {
-        this.responseCode = responseCode;
-    }
-
-    public boolean isSuccess() {
-        return success;
-    }
-
-    public void setSuccess(boolean success) {
-        this.success = success;
-    }
-}
